Address comments on PR #256
- Organize ``blob_data_models.py`` a directory below ``blob_transactions`` due to issues with circular dependencies.
